Result 型の名前空間を変更
