+++ src/packages/JourneyLogLinkType/UseCases/Get/GetUseCaseInterface.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace JourneyLogLinkType\UseCases\Get;
 
-use Support\Result\Result;
+use Support\ResultType\Result;
 
 interface GetUseCaseInterface
 {
