+++ src/tests/Unit/JourneyLog/Application/Get/GetInteractorTest.php
@@ -25,7 +25,7 @@ use Mockery;
 use Mockery\MockInterface;
 use PHPUnit\Framework\Attributes\Test;
 use Support\Domain\ValueObjects\OrderNo;
-use Support\Result\Result;
+use Support\ResultType\Result;
 use Tests\TestCase;
 
 class GetInteractorTest extends TestCase
